Add PlotTimeline in tslib
Fix lint

Fix type error
