+++ optuna_dashboard/ts/components/StudyDetail.tsx
@@ -16,9 +16,9 @@ import { Link, useParams } from "react-router-dom"
 import { useRecoilValue } from "recoil"
 
 import { TrialTable } from "@optuna/react"
-import * as Optuna from "@optuna/types"
 import { actionCreator } from "../action"
 import { useConstants } from "../constantsProvider"
+import { studyDetailToStudy } from "../graphUtil"
 import {
   reloadIntervalState,
   useStudyDetailValue,
@@ -62,19 +62,7 @@ export const StudyDetail: FC<{
   const reloadInterval = useRecoilValue<number>(reloadIntervalState)
   const studyName = useStudyName(studyId)
   const isPreferential = useStudyIsPreferential(studyId)
-  const study: Optuna.Study | null = studyDetail
-    ? {
-        id: studyDetail.id,
-        name: studyDetail.name,
-        directions: studyDetail.directions,
-        union_search_space: studyDetail.union_search_space,
-        intersection_search_space: studyDetail.intersection_search_space,
-        union_user_attrs: studyDetail.union_user_attrs,
-        datetime_start: studyDetail.datetime_start,
-        trials: studyDetail.trials,
-        metric_names: studyDetail.metric_names,
-      }
-    : null
+  const study = studyDetailToStudy(studyDetail)
 
   const title =
     studyName !== null ? `${studyName} (id=${studyId})` : `Study #${studyId}`
