+++ tslib/react/src/index.ts
@@ -6,6 +6,7 @@ export { PlotHistory } from "./components/PlotHistory"
 export { PlotImportance } from "./components/PlotImportance"
 export { PlotIntermediateValues } from "./components/PlotIntermediateValues"
 export { PlotSlice } from "./components/PlotSlice"
+export { PlotTimeline } from "./components/PlotTimeline"
 export { PlotParallelCoordinate } from "./components/PlotParallelCoordinate"
 export { TrialTable } from "./components/TrialTable"
 export { GraphContainer } from "./components/GraphContainer"
