+++ tslib/react/src/utils/graphUtil.ts
@@ -1,0 +1,18 @@
+import * as Optuna from "@optuna/types"
+
+export const makeHovertext = (trial: Optuna.Trial): string => {
+  return JSON.stringify(
+    {
+      number: trial.number,
+      values: trial.values,
+      params: trial.params.reduce<{
+        [key: string]: Optuna.CategoricalChoiceType
+      }>((obj, p) => {
+        obj[p.name] = p.param_external_value
+        return obj
+      }, {}),
+    },
+    undefined,
+    "  "
+  ).replace(/\n/g, "<br>")
+}
